Use precondition for defensive programming
Object#assert raises in development, but only reports an error in
production.

Object#precondition raises an error unless the condition is met.

Let's use this convention instead of raising errors manually. We have to
name them, think of their types and so on. With precondition, we need to
think about program execution and bad state, not error hierarchies.