Introduce invoices table
The invoices are strictly for orders with tickets. I won't hold any
line-item data as I can infer it all from the orders. Tickets are the
line items. I won't hold any business data for Neuvents and clients too,
as I can hard-code ours and infer the client's from Stripe. No amount
too, it's in the order.

I need to know the next available invoice number and that's it.
Everything else I can hard-code or infer.