Support invoice customer details from Stripe and manual data
I am about to issue my first "but I forgot to check Issue Invoice
checkbox invoice" so we need to support both inferring the customer data
from stripe and the invoice itself. I'll do it from the production
console now but I'll try to create an admin interface for this as well
if it happens more often.