Make possible local app ports configurable
If this deploy script is to be reused by other apps, the local app ports
cannot remain hardcoded to 3001/3002 as otherwise two apps deployed to
the same server will conflict when they attempt to listen on the same
ports. This commit makes them configurable, so multiple apps can be
configured to use different port ranges.