Don't stop unrelated containers during deploy
The deploy script assumed that the app's containers are the only ones
running on the server, so it could stop unrelated containers during
deploy. This generally makes the deploy script a bad citizen, and also
makes it impossible to reuse it to deploy multiple apps on the same
server.

This commit fixes that by filtering the containers that get manipulated
during deploy.