Don't delete unrelated images during deploy clean up
The deploy script assumed that the app's images are the only ones on the
server, so it could delete unrelated images in the clean up stage. This
generally makes the deploy script a bad citizen, and also makes it
impossible to reuse it to deploy multiple apps on the same server, as
their deploys would delete each other's images.

This commit fixes that by filtering the images that get deleted during
clean up.