Introduce Ruby Banitsa 2024, and basic multi-event support
Now that we are hosting multiple events at the same time, we need to
differentiate them somehow. For now, this happens on a per-host basis.

We still need more work: need to put speakers, talks, sponsors and
community partners in the database. Need to augment the events with
knowledge of the host, contact information, and more. This will come
when I have more time, for now the hacks reign supreme.