Prepare the admin interface for our multi-event tenancy world
We are running 2 simultaneous event at the moment: Ruby Banitsa 2024 and
Balkan Ruby 2025 on different hosts. Each has it admin and shows a
mixture of **ALL** orders and current event (chosen by host) metrics. We
need to clearly differentiate this in the admin and this is the first
step.

We are yet to choose different current events or create event specific
items, or configure events through the admin, but it is a step forward!

I have separated the orders in their own view, that shows them per
event. We need this for accounting. I have also introduced a new
navigation, a sidebar that we can use to fill with all the other admin
items we need to, and want to, configure through an interface in the
future.

What is missing is the actual event selection from that sidebar. This
will come later, as we have more business critical needs to focus on at
the moment.