Refactor admin tables into a single component
The `admin_table` view component creates table representations for our
admin needs. We have a very simple table definition and they all look
the same now. But! We will handle pagination and optional custom
representations when the time comes.

For now, I'm happy that we have a single table component component. I'm
not super happy with how the orders table looks but this is fixable.
With time, my child!