Introduce accepts_nested_attributes_for with destroy_missing: argument
Setting a checkbox for deleting a record is not a nice UX by the current
standards. I want to remove slots from the UX and for this to JUST WORK
when sending it to the backend. This is what the destroy_missing:
keyword argument does. It destroys the records that are missing in the
nested attributes setter.