Fix needles "more pages indication" (...) in 3 paged pagination
If we have 3 pages, we'll display:

[ < ][ 1 ][ ... ][ 2 ][ 3 ][ > ]

We don't need the 'more pages ...' indication for this case because we
can show all them pages. While we always want to display the first and
the last page for quick access, we need to guard for this case so we
don't show this needles indication.