Render HTML content as Markdown
I want to migrate us to use Markdown for rendering HTML content. So far
I have written straight HTML in the legacy YAML files and I have carried
this HTML over to the database. No more! I want to migrate this content
to Markdown while it's still not out of hand.

The first step is to start rendering it as Markdown. Redcarpet handles
HTML mixed in to the markdown so infront of it, the HTML content we have
is a valid Markdown. Good! Now I have to go over the content in the
database and ensure it is rendered well.