Use a single image gallery component
The Banitsa's style hero image split is nice when we have full control
over the arrangement, but now that we use the admin to build all future
admins, it doesn't make sense to:

a) build this kind of control in the admin
b) bother to implement the above

The current image gallery component we have based on snap scrolling
looks great as is. Delete!