Add a discount code field during checkout
We can use the discount field to pre-fill a stripe promotion code. We
are willing to be lenient here and still show the stripe promotion code
field if a non-valid discount code is applied so the user can try again.

A nice thing about the implementation is that we can have the discount
pre-applied in the form with the `discount_code=XXX` HTTP query
parameter.