Export invoices issued the last day of the month
Using a date will generate an SQL BETWEEN query that won't include
invoices issued in the last day:

```sql
SELECT \"orders\".*
FROM \"orders\"
WHERE (completed_at IS NOT NULL) AND \"orders\".\"completed_at\" BETWEEN '2025-03-01' AND '2025-03-31'
ORDER BY completed_at DESC
```

The query uses the BETWEEN operator with the range '2025-03-01' and
'2025-03-31'. In SQLite, when a date is provided as a string in the
format 'YYYY-MM-DD', it is interpreted as 'YYYY-MM-DD 00:00:00'
(midnight of that day)```

Therefore:

- The lower bound '2025-03-01' is interpreted as '2025-03-01 00:00:00'.
- The upper bound '2025-03-31' is interpreted as '2025-03-31 00:00:00'.

This means the query will include rows where completed_at is exactly at
midnight on March 31, 2025 ('2025-03-31 00:00:00'), but it will exclude
any rows where completed_at has a time component later in the day, such
as '2025-03-31 12:34:56'.

Using a time object for the date range will generate a berween operator
that includes the whole day like:

```sql
BETWEEN '2025-03-01 00:00:00' AND '2025-03-31 23:59:59'
```