Inline admin table actions to the right
I'm assuming that table data that has "inline-block" has an button
nested somewhere. If there are buttons, align those columns to the
right. The heuristic can be better if I had the admin button abstraction
in the CSS and not the Rails template helpers, but hey, I don't get to
change any of the existing tables. I'll get this for now as I am still
not sold that extra app-specific classes are a good idea in tailwind.